* riscos/soloader/dynamic/ld.so-1.9.9/d-link/riscos/resolve.s: Simplify by calling PLT fix up routine directly rather than through a function pointer.
* riscos/soloader/dynamic/ld.so-1.9.9/d-link/riscos/elfinterp.c:
Change visibility of symbol _dl_riscos_resolver to hidden. This
makes it visible to other compilation units in the library, but
not to other libraries and a PLT entry is not generated for it.
This is necessary to prevent the change to resolve.s above
causing an infinite loop in the dynamic linker.


git-svn-id: svn://svn.riscos.info/gccsdk/trunk@3507 5d7976f4-0b0e-0410-a586-d2bac8e3be52