Add game Neverball.
This port is targeted at the RaspberryPi and makes use of the
hardware accelerated OpenGL graphics via SDL2/Khronos.

At this time there is no sound support in SDL2 and it's full
screen only.

On the RaspberryPi 3, it seems to average about 40fps.

The packaged/zip'ed size of the game is about 71MB.


git-svn-id: svn://svn.riscos.info/gccsdk/trunk@7097 5d7976f4-0b0e-0410-a586-d2bac8e3be52