Update libreadline7 to create package and shared libs.
Updated libreadline7 setvars to create proper packages, including
a dev package and a shared libs package for bash to use.

Updated bash setvars to depend on SharedLibs and LibReadline7.
Note: this needs to be fixed if you want to build without shared
libraries, but it's better to be cautious with the dependencies.