ICU should be built as shared, looking for "*-*riscos*".
One common problem I've noticed with packages requiring attention
to be compiled with GCC 10 is that they're searching for a target
type with a pattern such as "arm-unknown-riscos*' or "*-*-riscos*",
which fails to match against the new "arm-riscos-gnueabihf" triplet.

Fortunately, it's easy to match both of them with "*-*riscos*".
Slightly trickier is the case when setvars needs to use the
correct prefix for the toolchain, or the qmake config file.
I'm going to try to patch the setvars file to work with either
RISC OS ABI, by testing if RO_USE_ARMEABIHF=yes.