Update vim to fix GCC 10 build failure and packaging.
GCC 10 complains by default if you define the same global variable
in multiple places. Add a patch to fix this for the version of vim
that's currently in Debian stable.

More importantly, I had to remove a spell check file containing
an "@" character from the archive because PackMan will fail to
extract it with a cryptic SWI error about being unable to create
a file. Apologies to any Serbian speakers in Serbia using the
Latin alphabet in RISC OS (the spell check file I had to delete).