Update Qt ports to version 5.5.1 and to use GCC 10.
Update the Qt 5 ports to 5.5.1 and fix the toolchain to use GCC 10.
With this change, Qt 5 autobuilds will require RO_USE_ARMEABIHF=yes.
I've turned off the VFP build, since GCC 10 assumes we have VFPv3.

I also patched Qt5Webkit to define __ARM_ARCH_7A__, because, when
compiling for an ARMv8-A target, __ARM_ARCH_8A__ is defined, which
is too new for this version of WebKit to recognize.