Upgrade develop/gcc from GCC 10.2.0 to 10.3.0.
I had a strange build failure with this version, which I found
a workaround for. When building libatomic with the cross-compiler,
but not the second time when it builds the native compiler, I
get a strange preprocessor error about a missing value for a
binary operand in libatomic/libatomic_i.h, due to IFUNC_NCOND(N)
apparently not expanding to a numeric value.

As a workaround, I added a patch to libatomic/libatomic_i.h to
replace the troublesome IFUNC_NCOND(N) with 0. Hopefully this
update fixes more bugs than it introduces.