Update x11/libxrender1 to a version that exists and compiles.
This build failed due to Debian no longer hosting that version.
I tried updating to 0.9.10, but it required a newer version of
X11 than we claim to support. 0.9.8 did compile and install.

Please note that the "preprocess" file applies the Debian patch and
also needed to be modified. I added a comment to the setvars
file to warn people about remembering to update them together.