Partial GCC 10 fixes for Raspberry Pi userland VPU driver.
Build the Khronos module using GCC 4, and build the shared libs
and apps using GCC 10, if configured (RO_USE_ARMEABIHF=yes).

Unfortunately, the demo apps built with GCC 10 crash with
"Fatal signal received: EMT trap" inside libunixlib on start.

I first tried adding "-fno-strict-aliasing", in case it was a
pointer aliasing optimization issue, but that didn't help
(although I suspect it may be needed because of the code
that converts between ints and floats using a union).

The LR values in the stack trace didn't seem to correspond
to functions within libunixlib.so, after subtracting the
library base address, and the backtrace printing code
couldn't figure it out either. I found the code in UnixLib
that prints the error and stack trace, in signal/post.c.
Enabling DEBUG in that file didn't provide any further info.

I think the crash may be due to missing support in UnixLib for
pthreads and EABI, as there's a message in its README-DEV file
stating that pthread support is currently broken for
non-chunked stack builds (#define __UNIXLIB_CHUNKED_STACK 0),
which is the case for EABI builds and not for ABI 2.0 builds.

I suspect the bug is in the code related to pthread stack setup,
growing the stack, or perhaps a bug in the EABISupport module.
The demo app doesn't create any threads, but the library code
that communicates with the VPU does create a notifier thread.