Initial work to add optional Ada language support to GCC 10.
Add a new RO_GCC_ADD_LANGUAGES variable that can be set to
"ada" in your build-setvars file to enable support for GNAT,
the Ada language compiler, standard library, and tools.

This feature doesn't work yet, because libgnat needs some
work to be ported to RISC OS, and the setvars file also needs
updating to include gnatmake and the other new tools in the
packages for the native compiler. I've added the first patch,
to gcc/ada/terminals.c, to not try to use POSIX ptys in libgnat.