Upgrade GCC 10.3.0 build to use binutils 2.31.1.
Upgrade GCC 10.3.0 to use binutils 2.31.1, merging the RISC OS
patches to apply cleanly to the updated logic in bfd/elf.c.

Note that the new binutils now requires autoconf 2.69, which is
newer than the autoconf that GCC 4 builds, so I've commented out
the lines in setvars to use the autoconf from GCC 4 binutils.
You'll now need to have autoconf 2.69 installed.