Upgrade the GCC 4 tools autoconf scripts to compatible versions.
Some of the autoconf/libtool script versions weren't compatible with
the versions of other scripts for the tools built by "make ronative".

These files should probably not be in the repo at all, but I wanted to
check in a working version of all of them in preparation to delete them
from the repo once I'm sure I can regenerate them all with the autoconf
built by the scripts for binutils and GCC 4.