Work in progress on LLVM/Clang 12.0.1 cross-compiler.
Create two new ports, develop/llvm12 and develop/clang12, and add
a new RISC OS type and toolchain definition. Both packages compile
and install, as cross-compilers only, but more work is needed to
tell Clang where to search for include files and linker objects,
e.g. crt0.o.

Clang also can't find its unit tests right now. The good news is
that only three of the LLVM unit tests are failing for me.

I've changed the target triplet to "armv7a-unknown-riscos-gnueabihf",
with the first part telling LLVM to generate code for ARMv7-A by
default, and the "-unknown-" to ensure that all of the components
are parsed correctly while working on the port. It should be possible
to switch to "armv7a-riscos-gnueabihf" once everything's working.

I had to add a CMake package (also not for RISC OS, yet), because
otherwise LLVM couldn't find the new CMake modules installed into
the env subdirectory by LLVM, and CMake fails. Building a current
version of CMake to install into the same location solves this.

I've patched gcc4/porting-scripts/ro-config and "build-program"
to support RO_USE_CLANG=yes, but I haven't been able to build
anything with the new toolchain until it can find crt0.o.

If you're curious about the possibility of building Clang and LLVM
as a native compiler and libraries, the first stumbling block
is a build failure due to missing std::wstring in our libstdc++.