Update README.md to explain how to customize GCC 10 build.
Update README.md to explain the use of the new RO_GCC_CONFIG_FLAGS
variable to customize the GCC 10.3.0 build to change the default
CPU to optimize for. I discovered that it's necessary to specify
the FPU as well as the CPU or else there's a build failure in
UnixLib because the assembler doesn't think we have VFP instructions.