Update the PDF build to work with GCC 10.
Several patches were required to fix build errors with GCC 10.
The declaration of a friend method can't include default arguments
unless it also includes a definition, so I removed the defaults and
explicitly specified those parameters. I also had to patch the
Makefiles to change references to "gcc" to use the cross-compiler.