Add "-fPIC" to Qt 5 static library rules to fix link errors.
Qt WebKit builds are failing when I build it using qmake, which
is necessary to enable all the WebKit features. The problem is
that the package links its subcomponents as static libraries
which it then links into the main shared library. Those files
all need to be compiled with "-fPIC" for the final library.

Modify the RISC OS qmake.conf files to behave like other ELF
platforms which add the shared library flags to CFLAGS when
building static libraries to avoid these types of errors.

I also added the font packages that Qt 5 requires to "depends"
to ensure that they get built to install with the package.