Fix Qt 5 WebKit build failures and configuration.
Qt WebKit builds were failing due to not being able to find
<QMediaPlayer>, because the include path wasn't being added.

Further investigation revealed that it's necessary to use
qmake to build this package instead of the "build-webkit" script,
in order for the generated Makefiles to enable all of the
available features and search the correct directories.
I also had to remove some files from the package that are
no longer being built.