Remove my hack to work around GCC compiler error with NEON.
Remove my hack to add a "-fno-peephole2" option to work around
a GCC internal error when NEON is used. Because NEON seems to
be unreliable in general, it's probably best not to use it by
default. Once the code generation errors have been fixed, the
internal compiler error should be fixable as well.