Update Qt 5.5.1 and Khronos in preparation for Qt 5.6.3.
In preparation to update the Qt port from 5.5.1 to Qt 5.6.3 LTS,
I've made a number of changes to the packages, most notably enabling
OpenGL (and adding a dependence on the Khronos package). This isn't
ideal for non-Raspberry Pi owners, but software rendering continues
to work, and OpenGL is required in order to build QML Quick, which
is required to support Qt WebEngine (the Blink-based replacement for
Qt WebKit, which was removed in Qt 5.6). Ultimately there should
be a common libEGL.so and libGLESv2.so for all RISC OS machines,
and apps shouldn't be linking against libbcm_host.so.

- Renamed the Qt OS type from "arm-riscos-g++" to "riscos-g++" to
match the other OS targets, which don't include the CPU. I've also
removed the VFP build rules in some cases, because VFP is the default
with the new GNUEABI, so a special VFP subdirectory has no purpose.
I also changed the Khronos port to not install its files into "vfp"
subdirectories, and to link its examples against the shared libs.

- Added new packages for Qt QML and Qt WebChannel, in preparation
for a port of Qt WebEngine after upgrading everything to Qt 5.6.3.

- Changed the Qt WebKit package to build using qmake instead of
the script it was using, which enabled many additional features
to be detected, especially after enabling OpenGL in Qt5Base.
Consequently, the package now depends on Qt5Qml and Qt5WebChannel.

- Removed patches from Qt5Base to add __attribute__((packed)) to
several data structures to save RAM. The explanation in the ReadMe
makes sense for QChar and QString, but fortunately these classes
were significantly optimized in Qt 5.6, making this optimization
unnecessary, so it makes sense to remove it now. The related patches
to add the packed attribute elsewhere seem even less likely to save
RAM and more likely to impair GCC's ability to generate optimal code,
so I removed them as well.

- Updated the Qt packages to include all of the installed shared
libraries and plugins in the RISC OS package. In addition, moved
the plugin locations to match other systems, as subdirectories of
"plugins" rather than "plugins/platforms". This also required
embedding the proper prefix into the app, instead of the prefix
that the cross-build installs to on the host machine.

- I also removed the qt.conf files generated by setvars for the
test apps, which set the Prefix to the app directory, instead of
the system-wide one. We'll also have to remove "Prefix=", or the
entire qt.conf file, from other apps that include a qt.conf that
overrides the Qt prefix, such as Otter browser.