Update Otter to create a Release build and remove qt.conf.
Configure the Otter browser as a release build, and remove the
qt.conf file that's setting the Qt prefix to the app directory.

I'm seeing Otter exit within about a minute with a message about
RISC OS not being able to move memory to increase the size of the
dynamic area that QDynamicArea is managing. This seems to be
more of a Qt bug than a Qt WebKit or Otter bug.