Check for RISC OS 5 sprite mode word support and use if available.
Instead of testing an environment variable in QSysInfo to determine
whether to swap the red and blue pixel components, move the
requireRedBlueSwap() static method to QRiscosBackingStore and
return false if we have support for RISC OS 5 sprite mode words,
which allow us to have the OS swap TBGR with the TRGB format that
Qt would prefer.

This also requires the use of OS_SpriteOp 52 (Paint Sprite Scaled)
instead of the faster OS_SpriteOp 34 (Put Sprite at User Coords).
My theory is that the OS will be able to do the conversion faster
than Qt can, but this should be verified.

The important part is getting RISC OS specific global state out of
the global Qt C++ classes, for easier merging of future Qt versions.