Upgrade Qt port from 5.5.1 to 5.6.3, adding Qt5Sensors.
Upgrade the Qt port to 5.6.3, merging the RISC OS patches and adding
necessary new ones. The "__attribute__((packed))" line that I added
was to fix a build assertion failure with GCC 4.7.4 (expecting the
size of the struct to be 1 byte, which it is with GCC 10), but my
successful testing has been with GCC 10.

I did update the Qt build scripts to build with GCC 4.7.4, but I
wasn't able to run any of the Qt demos that I built with GCC 4.7.4.
They crashed with "UnixLib detected recursion of SIGSEGV.  Exiting."