Add Debian 9 ("stretch") to repos to fix libssl1.0.2.
With the recent release of Debian 11, it becomes the new "stable",
and Debian 10 is the new "oldstable". Unfortunately, Qt 5.6 requires
OpenSSL 1.0, not 1.1, and the source files used for the libssl1.0.2
port could no longer be found.

As a workaround, until OpenSSL 1.0 is no longer needed, add "stretch"
to the list of sources to download from, after "stable" and "oldstable".