Add C11 functions in SharedCLibrary
This provides the following functions that were added in SharedCLibrary
6.05: aligned_alloc, at_quick_exit, quick_exit, timespec_get.

It also acknowledges the addition of the following functions in
SharedCLibrary 6.09, although all three of them were already supported as
statically linked functions in FPA builds: coshf, sinhf, tanhf.

The patch doesn't implement the functions in UnixLib builds, doesn't include
any of the atomic functions that were added in SharedCLibrary 6.14, and it
doesn't do any additional version checks, so it'll crash if the new
functions are called with an older version of SharedCLibrary.

Patch submitted by Cameron Cawley.


git-svn-id: svn://svn.riscos.info/gccsdk/trunk@7722 5d7976f4-0b0e-0410-a586-d2bac8e3be52