Revert to GCC 10.2.0 and binutils 2.30, with fixes.
Revert GCC version from 10.3.0 to 10.2.0, and binutils from 2.31.2
to 2.30.0, and remove my patches for custom default CPU/FPU flags.

Now the GCC and binutils versions match upstream Autobuilder, with
the exception that this is release 2 and includes the wchar_t type.

I've also fixed the gcc.dwarf2out.c.p patch to correctly apply, and
copied the "reconf-libstdc++" and "reconf-libunixlib" scripts locally,
so that I could modify them to always use autoconf2.69, not "autoconf".